Add IsMgmtErrorRef interface
Part of VRVDR-36175.  Add new IsMgmtErrorRef interface with private method implemented in base MgmtError object so we can check objects against this interface to know if they include a MgmtError object, and thus (for the purposes of this bug) should be encoded as a MgmtError when marshalled for JSON / RFC7951 etc.

Separate PR in configd to follow which uses this interface.